requirements.txt
faster-whisper>=1.0.0
srt>=3.5.0
google-generativeai>=0.8.0
